Move math_constants to util
