+++ libopenage/util/math_constants.h
